+++ src/test/java/com/caoccao/javet/values/reference/TestV8ValueProxy.java
@@ -21,14 +21,71 @@ import com.caoccao.javet.BaseTestJavetRuntime;
 import com.caoccao.javet.exceptions.JavetException;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.*;
 
 public class TestV8ValueProxy extends BaseTestJavetRuntime {
     @Test
-    public void testProxy() throws JavetException {
+    public void testEmptyHandler() throws JavetException {
+        try (V8ValueObject v8ValueObject = v8Runtime.getExecutor("const x = {a:1,b:2}; x;").execute()) {
+            try (V8ValueProxy v8ValueProxy = v8Runtime.createV8ValueProxy(v8ValueObject)) {
+                assertNotNull(v8ValueProxy);
+                assertFalse(v8ValueProxy.isRevoked());
+                assertEquals(1, v8ValueProxy.getInteger("a"));
+                assertEquals(2, v8ValueProxy.getInteger("b"));
+                assertTrue(v8ValueProxy.get("c").isUndefined());
+                v8ValueObject.set("c", 3);
+                assertEquals(3, v8ValueProxy.getInteger("c"));
+                v8ValueProxy.revoke();
+                assertTrue(v8ValueProxy.isRevoked());
+            }
+        }
+    }
+
+    @Test
+    public void testEmptyTargetAndEmptyHandler() throws JavetException {
+        try (V8ValueProxy v8ValueProxy = v8Runtime.createV8ValueProxy()) {
+            assertNotNull(v8ValueProxy);
+            assertFalse(v8ValueProxy.isRevoked());
+            try (IV8ValueObject iV8ValueObject = v8ValueProxy.getTarget()) {
+                try (IV8ValueArray iV8ValueArray = iV8ValueObject.getOwnPropertyNames()) {
+                    assertEquals(0, iV8ValueArray.getLength());
+                }
+                iV8ValueObject.set("a", 1);
+            }
+            assertEquals("{\"a\":1}", v8ValueProxy.toJsonString());
+        }
+    }
+
+    @Test
+    public void testHandlerGet() throws JavetException {
+        try (V8ValueObject v8ValueObject = v8Runtime.getExecutor("var x = {a:1,b:2}; x;").execute()) {
+            try (V8ValueProxy v8ValueProxy = v8Runtime.createV8ValueProxy(v8ValueObject)) {
+                assertNotNull(v8ValueProxy);
+                assertFalse(v8ValueProxy.isRevoked());
+                try (IV8ValueObject iV8ValueObjectHandler = v8ValueProxy.getHandler()) {
+                    iV8ValueObjectHandler.bindFunction("get", "(target, prop, receiver) => {" +
+                            "  if (prop in target) {" +
+                            "    return target[prop] + 1;" +
+                            "  }" +
+                            "  return undefined;" +
+                            "};");
+                }
+                assertEquals(2, v8ValueProxy.getInteger("a"));
+                assertEquals(3, v8ValueProxy.getInteger("b"));
+                v8ValueProxy.revoke();
+                assertTrue(v8ValueProxy.isRevoked());
+                assertTrue(v8ValueProxy.get("a").isUndefined());
+                assertTrue(v8ValueProxy.get("b").isUndefined());
+            }
+        }
+    }
+
+    @Test
+    public void testNativeProxyCreation() throws JavetException {
         try (V8ValueProxy v8ValueProxy = v8Runtime.getExecutor(
                 "const b = {}; const a = new Proxy(RegExp, b); a;").execute()) {
             assertNotNull(v8ValueProxy);
+            assertFalse(v8ValueProxy.isRevoked());
         }
     }
 }